Rename 'error' to 'jsonerror' to emphasise that these are JSON errors
