+++ src/github.com/matrix-org/dendrite/clientapi/jsonerror/jsonerror.go
@@ -1,4 +1,4 @@
-package error
+package jsonerror
 
 import "fmt"
