+++ src/github.com/matrix-org/dendrite/clientapi/jsonerror/jsonerror_test.go
@@ -1,4 +1,4 @@
-package error
+package jsonerror
 
 import (
 	"encoding/json"
